# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AssetCatalogResponse": {
        "additionalProperties": false,
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/ProductAsset"
            },
            "maxItems": 11,
            "minItems": 11,
            "title": "Assets",
            "type": "array"
          },
          "disclaimer": {
            "title": "Disclaimer",
            "type": "string"
          },
          "domain": {
            "const": "oil_gas_transfer",
            "title": "Domain",
            "type": "string"
          },
          "educational_only": {
            "const": true,
            "title": "Educational Only",
            "type": "boolean"
          },
          "profile_id": {
            "const": "otsoc.asset_inventory.oil_gas_transfer",
            "title": "Profile Id",
            "type": "string"
          },
          "profile_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Profile Sha256",
            "type": "string"
          },
          "profile_version": {
            "const": "1.0.0",
            "title": "Profile Version",
            "type": "string"
          },
          "relationships": {
            "items": {
              "$ref": "#/components/schemas/RelationshipDefinition"
            },
            "maxItems": 9,
            "minItems": 9,
            "title": "Relationships",
            "type": "array"
          },
          "zones": {
            "items": {
              "$ref": "#/components/schemas/ZoneDefinition"
            },
            "maxItems": 5,
            "minItems": 5,
            "title": "Zones",
            "type": "array"
          }
        },
        "required": [
          "profile_id",
          "profile_version",
          "profile_sha256",
          "domain",
          "educational_only",
          "disclaimer",
          "zones",
          "assets",
          "relationships"
        ],
        "title": "AssetCatalogResponse",
        "type": "object"
      },
...
      "AssetDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "asset": {
            "$ref": "#/components/schemas/ProductAsset"
          },
          "inbound_relationships": {
            "items": {
              "$ref": "#/components/schemas/RelationshipDefinition"
            },
            "title": "Inbound Relationships",
            "type": "array"
          },
          "outbound_relationships": {
            "items": {
              "$ref": "#/components/schemas/RelationshipDefinition"
            },
            "title": "Outbound Relationships",
            "type": "array"
          },
          "profile_id": {
            "const": "otsoc.asset_inventory.oil_gas_transfer",
            "title": "Profile Id",
            "type": "string"
          },
          "profile_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Profile Sha256",
            "type": "string"
          },
          "profile_version": {
            "const": "1.0.0",
            "title": "Profile Version",
            "type": "string"
          },
          "zone": {
            "$ref": "#/components/schemas/ZoneDefinition"
          }
        },
        "required": [
          "profile_id",
          "profile_version",
          "profile_sha256",
          "asset",
          "zone",
          "inbound_relationships",
          "outbound_relationships"
        ],
        "title": "AssetDetailResponse",
        "type": "object"
      },
...
      "AssetOverviewSummary": {
        "additionalProperties": false,
        "properties": {
          "cyber": {
            "const": 6,
            "title": "Cyber",
            "type": "integer"
          },
          "enabled": {
            "maximum": 11.0,
            "minimum": 0.0,
            "title": "Enabled",
            "type": "integer"
          },
          "process": {
            "const": 5,
            "title": "Process",
            "type": "integer"
          },
          "total": {
            "const": 11,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "enabled",
          "cyber",
          "process"
        ],
        "title": "AssetOverviewSummary",
        "type": "object"
      },
...
      "CorrelationOverviewSummary": {
        "additionalProperties": false,
        "properties": {
          "correlated": {
            "minimum": 0.0,
            "title": "Correlated",
            "type": "integer"
          },
          "indeterminate": {
            "minimum": 0.0,
            "title": "Indeterminate",
            "type": "integer"
          },
          "insufficient_evidence": {
            "minimum": 0.0,
            "title": "Insufficient Evidence",
            "type": "integer"
          },
          "not_correlated": {
            "minimum": 0.0,
            "title": "Not Correlated",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "correlated",
          "not_correlated",
          "insufficient_evidence",
          "indeterminate"
        ],
        "title": "CorrelationOverviewSummary",
        "type": "object"
      },
...
      "EvidenceRecordResponse": {
        "additionalProperties": false,
        "properties": {
          "canonical_byte_length": {
            "title": "Canonical Byte Length",
            "type": "integer"
          },
          "evidence_id": {
            "format": "uuid",
            "title": "Evidence Id",
            "type": "string"
          },
          "evidence_type": {
            "title": "Evidence Type",
            "type": "string"
          },
          "evidence_version": {
            "title": "Evidence Version",
            "type": "integer"
          },
          "integrity_sha256": {
            "title": "Integrity Sha256",
            "type": "string"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HistoricalCoolingTelemetryPayloadV1"
              },
              {
                "$ref": "#/components/schemas/OilGasTelemetryPayloadV2"
              },
              {
                "$ref": "#/components/schemas/SyntheticModbusEvent"
              },
              {
                "$ref": "#/components/schemas/ProtocolSemanticEvent"
              },
              {
                "$ref": "#/components/schemas/AssetContextEvent"
              },
              {
                "$ref": "#/components/schemas/CommunicationPolicyFinding"
              },
              {
                "$ref": "#/components/schemas/CyberPhysicalCorrelationFinding"
              }
            ],
            "title": "Payload"
          },
          "payload_schema": {
            "title": "Payload Schema",
            "type": "string"
          },
          "payload_schema_version": {
            "enum": [
              "1.0.0",
              "2.0.0"
            ],
            "title": "Payload Schema Version",
            "type": "string"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HistoricalEvidenceProvenanceV1"
              },
              {
                "$ref": "#/components/schemas/EvidenceProvenance"
              },
              {
                "$ref": "#/components/schemas/SyntheticProtocolProvenance"
              },
              {
                "$ref": "#/components/schemas/SemanticDerivationProvenance"
              },
              {
                "$ref": "#/components/schemas/AssetContextDerivationProvenance"
              },
              {
                "$ref": "#/components/schemas/PolicyFindingDerivationProvenance"
              },
              {
                "$ref": "#/components/schemas/CorrelationDerivationProvenance"
              }
            ],
            "title": "Provenance"
          },
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          },
          "sequence_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sequence Number"
          },
          "source_event_id": {
            "title": "Source Event Id",
            "type": "string"
          },
          "source_key": {
            "title": "Source Key",
            "type": "string"
          }
        },
        "required": [
          "evidence_id",
          "evidence_version",
          "source_key",
          "source_event_id",
          "evidence_type",
          "observed_at",
          "received_at",
          "sequence_number",
          "payload_schema",
          "payload_schema_version",
          "payload",
          "provenance",
          "integrity_sha256",
          "canonical_byte_length"
        ],
        "title": "EvidenceRecordResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IncidentOverviewSummary": {
        "additionalProperties": false,
        "properties": {
          "categories": {
            "$ref": "#/components/schemas/IncidentCategorySummary"
          },
          "high": {
            "minimum": 0.0,
            "title": "High",
            "type": "integer"
          },
          "high_non_resolved": {
            "minimum": 0.0,
            "title": "High Non Resolved",
            "type": "integer"
          },
          "investigating": {
            "minimum": 0.0,
            "title": "Investigating",
            "type": "integer"
          },
          "low": {
            "minimum": 0.0,
            "title": "Low",
            "type": "integer"
          },
          "medium": {
            "minimum": 0.0,
            "title": "Medium",
            "type": "integer"
          },
          "open": {
            "minimum": 0.0,
            "title": "Open",
            "type": "integer"
          },
          "resolved": {
            "minimum": 0.0,
            "title": "Resolved",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "open",
          "investigating",
          "resolved",
          "low",
          "medium",
          "high",
          "high_non_resolved",
          "categories"
        ],
        "title": "IncidentOverviewSummary",
        "type": "object"
      },
...
      "OverviewRunContext": {
        "additionalProperties": false,
        "properties": {
          "configuration_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configuration Hash"
          },
          "context_scope": {
            "const": "CURRENT_RUN",
            "title": "Context Scope",
            "type": "string"
          },
          "evidence_simulation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Simulation Id"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "scenario_id": {
            "enum": [
              "BASELINE",
              "S1",
              "S2",
              "S3",
              "S4"
            ],
            "title": "Scenario Id",
            "type": "string"
          },
          "scenario_state": {
            "const": "COMPLETED",
            "title": "Scenario State",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "scenario_id",
          "scenario_state",
          "context_scope",
          "evidence_simulation_id",
          "configuration_hash"
        ],
        "title": "OverviewRunContext",
        "type": "object"
      },
      "OverviewSummaryResponse": {
        "additionalProperties": false,
        "properties": {
          "active_run": {
            "$ref": "#/components/schemas/OverviewRunContext"
          },
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "assets": {
            "$ref": "#/components/schemas/AssetOverviewSummary"
          },
          "correlations": {
            "$ref": "#/components/schemas/CorrelationOverviewSummary"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "incidents": {
            "$ref": "#/components/schemas/IncidentOverviewSummary"
          },
          "linked_valve_command": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceRecordResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "policy_findings": {
            "$ref": "#/components/schemas/PolicyOverviewSummary"
          },
          "process_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceRecordResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "process_snapshot_message": {
            "title": "Process Snapshot Message",
            "type": "string"
          },
          "process_snapshot_scope": {
            "enum": [
              "ACTIVE_RUN",
              "BASELINE_REFERENCE",
              "UNAVAILABLE"
            ],
            "title": "Process Snapshot Scope",
            "type": "string"
          },
          "process_snapshot_status": {
            "enum": [
              "COMPLETE",
              "UNAVAILABLE"
            ],
            "title": "Process Snapshot Status",
            "type": "string"
          },
          "recent_activity": {
            "items": {
              "$ref": "#/components/schemas/RecentActivity"
            },
            "title": "Recent Activity",
            "type": "array"
          },
          "window_complete": {
            "const": true,
            "title": "Window Complete",
            "type": "boolean"
          },
          "window_end": {
            "format": "date-time",
            "title": "Window End",
            "type": "string"
          },
          "window_start": {
            "format": "date-time",
            "title": "Window Start",
            "type": "string"
          }
        },
        "required": [
          "generated_at",
          "as_of",
          "window_start",
          "window_end",
          "window_complete",
          "active_run",
          "incidents",
          "policy_findings",
          "correlations",
          "assets",
          "recent_activity",
          "process_snapshot_status",
          "process_snapshot_scope",
          "process_snapshot_message",
          "process_snapshot",
          "linked_valve_command"
        ],
        "title": "OverviewSummaryResponse",
        "type": "object"
      },
...
      "PolicyOverviewSummary": {
        "additionalProperties": false,
        "properties": {
          "approved": {
            "minimum": 0.0,
            "title": "Approved",
            "type": "integer"
          },
          "denied": {
            "minimum": 0.0,
            "title": "Denied",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          },
          "unknown": {
            "minimum": 0.0,
            "title": "Unknown",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "approved",
          "denied",
          "unknown"
        ],
        "title": "PolicyOverviewSummary",
        "type": "object"
      },
...
      "ProductAsset": {
        "additionalProperties": false,
        "properties": {
          "asset_id": {
            "format": "uuid",
            "title": "Asset Id",
            "type": "string"
          },
          "definition": {
            "$ref": "#/components/schemas/AssetDefinition"
          },
          "process_point_ids": {
            "items": {
              "type": "string"
            },
            "title": "Process Point Ids",
            "type": "array"
          }
        },
        "required": [
          "asset_id",
          "definition",
          "process_point_ids"
        ],
        "title": "ProductAsset",
        "type": "object"
      },
...
      "RelationshipDefinition": {
        "additionalProperties": false,
        "properties": {
          "relationship_type": {
            "$ref": "#/components/schemas/RelationshipType"
          },
          "source_asset_key": {
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Source Asset Key",
            "type": "string"
          },
          "target_kind": {
            "$ref": "#/components/schemas/RelationshipTargetKind"
          },
          "target_ref": {
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Target Ref",
            "type": "string"
          }
        },
        "required": [
          "relationship_type",
          "source_asset_key",
          "target_kind",
          "target_ref"
        ],
        "title": "RelationshipDefinition",
        "type": "object"
      },
...
      "ReplayBundleResponse": {
        "additionalProperties": false,
        "properties": {
          "completeness": {
            "enum": [
              "COMPLETE",
              "PARTIAL"
            ],
            "title": "Completeness",
            "type": "string"
          },
          "configuration_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configuration Hash"
          },
          "correlation_evidence_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Evidence Id"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ReplayEvent"
            },
            "maxItems": 2000,
            "title": "Events",
            "type": "array"
          },
          "gaps": {
            "items": {
              "type": "string"
            },
            "title": "Gaps",
            "type": "array"
          },
          "incident": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IncidentRecordResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "lab_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lab Run Id"
          },
          "observed_from": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed From"
          },
          "observed_to": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed To"
          },
          "scenario_id": {
            "anyOf": [
              {
                "enum": [
                  "BASELINE",
                  "S1",
                  "S2",
                  "S3",
                  "S4"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "simulation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Simulation Id"
          },
          "source_kind": {
            "enum": [
              "INCIDENT",
              "CORRELATION",
              "EVIDENCE_WINDOW"
            ],
            "title": "Source Kind",
            "type": "string"
          },
          "truncated": {
            "const": false,
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "source_kind",
          "incident",
          "correlation_evidence_id",
          "simulation_id",
          "configuration_hash",
          "observed_from",
          "observed_to",
          "events",
          "completeness",
          "gaps",
          "truncated"
        ],
        "title": "ReplayBundleResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ZoneDefinition": {
        "additionalProperties": false,
        "properties": {
          "allowed_relationship_types": {
            "items": {
              "$ref": "#/components/schemas/RelationshipType"
            },
            "title": "Allowed Relationship Types",
            "type": "array"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "purpose": {
            "maxLength": 180,
            "minLength": 1,
            "title": "Purpose",
            "type": "string"
          },
          "trust_classification": {
            "$ref": "#/components/schemas/TrustClassification"
          },
          "zone_id": {
            "$ref": "#/components/schemas/ZoneId"
          }
        },
        "required": [
          "zone_id",
          "name",
          "purpose",
          "trust_classification",
          "allowed_relationship_types"
        ],
        "title": "ZoneDefinition",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OTSOCSessionCookie": {
        "description": "Opaque HttpOnly local session cookie. The configured runtime name may differ.",
        "in": "cookie",
        "name": "otsoc_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Repository and contract foundation for the isolated, advisory-only OT-SOC Fusion X academic prototype.",
    "title": "OT-SOC Fusion X",
    "version": "1.1.1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/assets": {
      "get": {
        "operationId": "read_asset_catalog_api_v1_assets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetCatalogResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Read Asset Catalog",
        "tags": [
          "product"
        ]
      }
    },
    "/api/v1/assets/{asset_key}": {
      "get": {
        "operationId": "read_asset_detail_api_v1_assets__asset_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_key",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
              "title": "Asset Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Read Asset Detail",
        "tags": [
          "product"
        ]
      }
    },
...
    "/api/v1/overview/summary": {
      "get": {
        "operationId": "read_overview_summary_api_v1_overview_summary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Read Overview Summary",
        "tags": [
          "product"
        ]
      }
    },
    "/api/v1/replay": {
      "get": {
        "operationId": "read_replay_bundle_api_v1_replay_get",
        "parameters": [
          {
            "in": "query",
            "name": "incident_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Incident Id"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "in": "query",
            "name": "correlation_evidence_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Correlation Evidence Id"
            }
          },
          {
            "in": "query",
            "name": "simulation_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "minLength": 5,
                  "pattern": "^[a-zA-Z0-9._:-]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Simulation Id"
            }
          },
          {
            "in": "query",
            "name": "configuration_hash",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[0-9a-f]{64}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Configuration Hash"
            }
          },
          {
            "in": "query",
            "name": "observed_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Observed From"
            }
          },
          {
            "in": "query",
            "name": "observed_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Observed To"
            }
          },
          {
            "in": "query",
            "name": "evidence_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Evidence Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayBundleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Read Replay Bundle",
        "tags": [
          "product"
        ]
      }
    },
...
  }
}